context 파일을 useContext 파일에 병합
