--- src/Contexts/AuthContext.ts
@@ -1,17 +1,0 @@
-import { createContext } from "react";
-
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-}
-
-interface AuthContextType {
-  user: User | null;
-  login: (email: string, password: string) => void;
-  logout: () => void;
-}
-
-export const AuthContext = createContext<AuthContextType | undefined>(
-  undefined,
-);
