--- src/Contexts/NotificationContext.ts
@@ -1,17 +1,0 @@
-import { createContext } from "react";
-
-export interface Notification {
-  id: number;
-  message: string;
-  type: "info" | "success" | "warning" | "error";
-}
-
-interface NotificationContextType {
-  notifications: Notification[];
-  addNotification: (message: string, type: Notification["type"]) => void;
-  removeNotification: (id: number) => void;
-}
-
-export const NotificationContext = createContext<
-  NotificationContextType | undefined
->(undefined);
