--- src/Contexts/ThemeContext.ts
@@ -1,10 +1,0 @@
-import { createContext } from "react";
-
-interface ThemeContextType {
-  theme: string;
-  toggleTheme: () => void;
-}
-
-export const ThemeContext = createContext<ThemeContextType | undefined>(
-  undefined,
-);
