+++ src/hooks/useAuthContext.tsx
@@ -1,5 +1,21 @@
 import { useContext } from "react";
-import { AuthContext } from "../Contexts/AuthContext";
+import { createContext } from "react";
+
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+interface AuthContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined,
+);
 
 export const useAuthContext = () => {
   const context = useContext(AuthContext);
