+++ src/hooks/useNotificationContext.tsx
@@ -1,5 +1,21 @@
 import { useContext } from "react";
-import { NotificationContext } from "../Contexts/NotificationContext";
+import { createContext } from "react";
+
+export interface Notification {
+  id: number;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+}
+
+interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
 
 export const useNotificationContext = () => {
   const context = useContext(NotificationContext);
