+++ src/hooks/useThemeContext.tsx
@@ -1,5 +1,13 @@
-import { useContext } from "react";
-import { ThemeContext } from "../Contexts/ThemeContext";
+import { createContext, useContext } from "react";
+
+interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
 
 export const useThemeContext = () => {
   const context = useContext(ThemeContext);
